Claude Version 0.02
This version adds Project Assignment 1 to considerations and updates the src/App.js file to be functional (removes syntax errors).

It combines the student/non-student/faculty classes into 1 user called Buyer and adds a text box for entering a password.  CofC community members get a discount, but it appears that the accounts currently have no way to distinguish student/faculty and nonstudent, so everyone is getting the discount. 

The app attempts to address special accomodations for Faculty/Staff seating access and Handicap access by requiring someone to click a check box when signing in, so this needs to be adjusted to associate with the account. Generally, accounts should need some sort of registration. 

Databases and backend still need implementation. Pay currently has no redirection to a third party payment system before simply giving the user the ticket. Transfer and return don't work.